Moved dashboard logic into query submodule
